Json refactoring for compilation fixes
