+++ include/stencil/WebService.h
@@ -1,4 +1,4 @@
-#include "DataHandlerJson.h"
+#include "jsonserdes.h"
 
 #pragma warning(push, 3)
 #pragma warning(disable : 4365)    // conversion from 'const char' to 'unsigned char', signed/unsigned mismatch
